processing payment

+++ payment-service/src/main/java/com/hrajaona/orderandpay/paymentservice/adapter/in/web/dto/PaymentRequest.java
@@ -4,12 +4,13 @@ import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Positive;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 public record PaymentRequest(
         @NotNull UUID orderId,
         @NotNull UUID userId,
-        @NotNull @Positive double amount,
+        @NotNull @Positive BigDecimal amount,
         @NotBlank String method,
         @NotNull String status,
         @NotBlank String reference) {
